sprint3_bug_fix, second attempt
