--- src/main/java/Login.java
@@ -1,13 +1,0 @@
-public class Login {
-    private String login;
-    private String password;
-
-
-    public Login(String login, String password) {
-        this.login = login;
-        this.password = password;
-
-    }
-
-
-}
