+++ src/main/java/ru/praktikum/CreateCourier.java
@@ -1,3 +1,5 @@
+package ru.praktikum;
+
 public class CreateCourier {
     private String login;
     private String password;
@@ -8,5 +10,8 @@ public class CreateCourier {
         this.password = password;
         this.firstName = firstName;
     }
+
+    public static final CreateCourier CREATE_COURIER = new CreateCourier("anna_ivanova","12345678","Anna");
+    public static final CreateCourier CREATE_COURIER_WITHOUT_PASSWORD = new CreateCourier("anna_14",null,"Анна");
 
 }
