+++ src/main/java/ru/praktikum/EndPoints.java
@@ -1,0 +1,12 @@
+package ru.praktikum;
+
+
+public class EndPoints {
+
+    public static final String ORDERS = "api/v1/orders";
+    public static final String COURIER = "api/v1/courier/";
+    public static final String LOGIN = "api/v1/courier/login" ;
+    public static final String BASE_URI = "http://qa-scooter.praktikum-services.ru/";
+
+
+}
