+++ src/main/java/ru/praktikum/Track.java
@@ -1,3 +1,5 @@
+package ru.praktikum;
+
 public class Track {
     public Integer track;
     public Integer getTrack() {
