+++ src/main/java/ru/praktikum/UserId.java
@@ -1,3 +1,5 @@
+package ru.praktikum;
+
 public class UserId {
     public Integer id;
     public Integer getId() {
